requirements.txt
# Apple Photos Export Tool - Python Dependencies
# Standard libraries for image processing and metadata extraction

# Core image processing
Pillow>=10.0.0          # EXIF data extraction and image processing
pillow-heif>=0.13.0     # HEIC/HEIF format support for Pillow
lxml>=4.9.0             # XML parsing for XMP files
python-dateutil>=2.8.0  # Advanced date parsing and manipulation

# Logging and configuration
loguru>=0.7.0           # Modern logging with colors and structured output
tqdm>=4.65.0
psutil>=5.9.0            # System and process monitoring

# Testing framework
pytest>=7.4.0           # Testing framework
pytest-cov>=4.1.0       # Coverage reporting
pytest-mock>=3.11.0     # Mocking utilities

# Development tools
black>=23.0.0           # Code formatting
flake8>=6.0.0           # Linting
mypy>=1.5.0             # Type checking

# File system utilities
pathlib2>=2.3.7         # Enhanced path operations (Python 3.4+ compatibility)
